requirements.txt
# Web Scraping
playwright==1.41.0
beautifulsoup4==4.12.3
lxml==5.1.0
feedparser==6.0.11
httpx>=0.27.1

# LLM Integration
anthropic==0.40.0

# Data Validation & Configuration
pydantic>=2.11.0
pydantic-settings>=2.7.0
pyyaml==6.0.2
python-dotenv==1.0.1

# Async & Utilities
aiofiles==24.1.0
tenacity==9.0.0

# Logging
loguru==0.7.3

# MCP Client (for Notion and Slack integration)
mcp==1.25.0
